requirements.txt
fastapi
uvicorn
pydantic
numpy
pillow
gymnasium
stable-baselines3
